Fix FMS pylon. Close #506
