+++ addons/uh60_fms/config/cfgVehicles.hpp
@@ -1,4 +1,5 @@
 #define FMS_PAGE_INDEX FMS_R_PAGE_INDEX
+#define FMS_PYLONS_0
 class VTX_FMS_R
 {
   topLeft="FMS_ScreenR_LH";
@@ -6,8 +7,6 @@ class VTX_FMS_R
   bottomLeft="FMS_ScreenR_LD";
   #include "MFD\FMS.hpp"
 }; // FMS_R
-#undef FMS_PAGE_INDEX
-#define FMS_PAGE_INDEX FMS_L_PAGE_INDEX
 class VTX_FMS_L
 {
   topLeft="FMS_ScreenL_LH";
@@ -15,6 +14,43 @@ class VTX_FMS_L
   bottomLeft="FMS_ScreenL_LD";
   #include "MFD\FMS.hpp"
 }; // FMS_L
+#undef FMS_PYLONS_0
+
+#define FMS_PYLONS_2
+class VTX_FMS_R2
+{
+  topLeft="FMS_ScreenR_LH";
+  topRight="FMS_ScreenR_RH";
+  bottomLeft="FMS_ScreenR_LD";
+  #include "MFD\FMS.hpp"
+}; // FMS_R2
+class VTX_FMS_L2
+{
+  topLeft="FMS_ScreenL_LH";
+  topRight="FMS_ScreenL_RH";
+  bottomLeft="FMS_ScreenL_LD";
+  #include "MFD\FMS.hpp"
+}; // FMS_L2
+#undef FMS_PYLONS_2
+
+#define FMS_PYLONS_4
+class VTX_FMS_R4
+{
+  topLeft="FMS_ScreenR_LH";
+  topRight="FMS_ScreenR_RH";
+  bottomLeft="FMS_ScreenR_LD";
+  #include "MFD\FMS.hpp"
+}; // FMS_R4
+#undef FMS_PAGE_INDEX
+#define FMS_PAGE_INDEX FMS_L_PAGE_INDEX
+class VTX_FMS_L4
+{
+  topLeft="FMS_ScreenL_LH";
+  topRight="FMS_ScreenL_RH";
+  bottomLeft="FMS_ScreenL_LD";
+  #include "MFD\FMS.hpp"
+}; // FMS_L4
+#undef FMS_PYLONS_4
 
 
 class CfgVehicles {
