+++ addons/uh60_mfd/config/MFD/pages/flir/laserinfo.hpp
@@ -14,5 +14,5 @@ class Lasers {
 	BOXEDTEXT_CONDITION(LSTE,0.09,0.16+0.02,0.07,0.04,"LST E",COND_ISNUMBER(USERMFDV_LST,4))
 	BOXEDTEXT_CONDITION(LSTF,0.09,0.16+0.02,0.07,0.04,"LST F",COND_ISNUMBER(USERMFDV_LST,5))
 
-	BOXEDTEXT_CONDITION(REMOTE,0.15,(BEZEL_Y6),0.1,0.04,"REMOTE","(1 - (((getUserMFDValue _this) select USERMFDV_LST) == ((getUserMFDValue _this) select USERMFDV_PRI_CH)))")
+	BOXEDTEXT_CONDITION(REMOTE,0.15,(BEZEL_Y6),0.1,0.04,"REMOTE","1 - (user94 <= user92) * (user94 >= user92)")
 };
